#454 BkReuse removed. Tests from BkReuseTest moved to BkBasicTest
