--- src/main/java/org/takes/http/BkReuse.java
@@ -1,51 +1,0 @@
-/**
- * The MIT License (MIT)
- *
- * Copyright (c) 2015 Yegor Bugayenko
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.takes.http;
-
-import java.io.IOException;
-import java.net.Socket;
-
-/**
- * Reusable back-end.
- *
- * @author Piotr Pradzynski (prondzyn@gmail.com)
- * @version $Id$
- */
-public final class BkReuse extends BkWrap {
-
-    /**
-     * Constructor of BkReuse.
-     * @param back Origin back-end.
-     */
-    @SuppressWarnings("PMD.UnusedFormalParameter")
-    public BkReuse(final Back back) {
-        super(new Back() {
-            @Override
-            public void accept(final Socket socket) throws IOException {
-                throw new UnsupportedOperationException();
-            }
-        });
-    }
-
-}
